api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
azure-functions==1.21.3
azure-data-tables==12.5.0
azure-monitor-query==1.4.0
azure-identity==1.19.0
msgraph-sdk==1.16.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
structlog==24.4.0
slowapi==0.1.9
